refactor: Extract ErrorCallbacks into its own module
